requirements.txt
gradio
google-generativeai
fuzzywuzzy
python-dotenv